requirements.txt
langchain==0.1.20
langchain-openai
python-telegram-bot==20.6
python-dotenv
